Refactoring for react auth
