+++ app/javascript/pages/SignIn.tsx
@@ -3,9 +3,10 @@ import React from "react"
 import { SignInForm } from "../authentication/SignInForm"
 import { LoggedInUserIdleTimeout } from "../authentication/timeout"
 import { UserProvider } from "../authentication/UserContext"
-import FormsLayout from "../layouts/Forms"
+import { withAppSetup } from "../layouts/app_settings"
+import FormsLayout from "../layouts/FormLayout"
 
-const SignIn = () => {
+const component = () => {
   return (
     <FormsLayout>
       <UserProvider>
@@ -15,5 +16,7 @@ const SignIn = () => {
     </FormsLayout>
   )
 }
+
+const SignIn = withAppSetup(component)
 
 export { SignIn as default, SignIn }
